Split exhibition.main into separate files
Fixes #49
