+++ exhibition/__init__.py
@@ -1,5 +1,5 @@
-__version__ = "0.0.3"
-
 from ._version import get_versions
+
 __version__ = get_versions()['version']
+
 del get_versions
