+++ exhibition/tests/test_command.py
@@ -24,12 +24,12 @@ import logging
 
 from click.testing import CliRunner
 
-from exhibition import command, main
+from exhibition import command, config
 
 
 class CommandTestCase(TestCase):
-    @mock.patch("exhibition.command.main.gen")
-    @mock.patch("exhibition.command.main.Config.from_path", return_value=main.Config())
+    @mock.patch("exhibition.command.utils.gen")
+    @mock.patch("exhibition.command.config.Config.from_path", return_value=config.Config())
     def test_gen(self, config_mock, gen_mock):
         runner = CliRunner()
         result = runner.invoke(command.exhibition, ["gen"])
@@ -38,10 +38,10 @@ class CommandTestCase(TestCase):
         self.assertEqual(gen_mock.call_count, 1)
         self.assertEqual(gen_mock.call_args, ((config_mock.return_value,), {}))
 
-        self.assertEqual(config_mock.call_args, ((main.SITE_YAML_PATH,), {}))
+        self.assertEqual(config_mock.call_args, ((config.SITE_YAML_PATH,), {}))
 
-    @mock.patch("exhibition.command.main.serve", return_value=(mock.Mock(), mock.Mock()))
-    @mock.patch("exhibition.command.main.Config.from_path", return_value=main.Config())
+    @mock.patch("exhibition.command.utils.serve", return_value=(mock.Mock(), mock.Mock()))
+    @mock.patch("exhibition.command.config.Config.from_path", return_value=config.Config())
     def test_serve(self, config_mock, serve_mock):
         runner = CliRunner()
         result = runner.invoke(command.exhibition, ["serve"])
@@ -50,7 +50,7 @@ class CommandTestCase(TestCase):
         self.assertEqual(serve_mock.call_count, 1)
         self.assertEqual(serve_mock.call_args, ((config_mock.return_value,), {}))
 
-        self.assertEqual(config_mock.call_args, ((main.SITE_YAML_PATH,), {}))
+        self.assertEqual(config_mock.call_args, ((config.SITE_YAML_PATH,), {}))
         self.assertEqual(serve_mock.return_value[0].shutdown.call_count, 0)
         self.assertEqual(serve_mock.return_value[1].join.call_count, 1)
 
@@ -61,7 +61,7 @@ class CommandTestCase(TestCase):
         self.assertEqual(serve_mock.call_count, 2)
         self.assertEqual(serve_mock.call_args, ((config_mock.return_value,), {}))
 
-        self.assertEqual(config_mock.call_args, ((main.SITE_YAML_PATH,), {}))
+        self.assertEqual(config_mock.call_args, ((config.SITE_YAML_PATH,), {}))
         self.assertEqual(serve_mock.return_value[0].shutdown.call_count, 1)
         self.assertEqual(serve_mock.return_value[1].join.call_count, 2)
